refactor(#1464): campsite split of python database.rs into database/ submodules
Pure code motion (epic #1116): bindings/python/src/database.rs (994L) becomes
bindings/python/src/database/{mod,open,write_methods}.rs. lib.rs is
byte-unchanged: `database::{open, Database}` and `register_database` still
resolve through mod.rs.

write_methods.rs takes the write-engine accessors (`writable_engine`,
`with_write_engine` — the issue's `require_writable` since renamed), the
flush_run/maintenance_step/write_stats pymethods and the private `execute_dml`
DML route (now `pub(super)`, since `execute()` stayed in mod.rs).

The open pyfunction lives in database/open.rs but is declared as
`#[path = "open.rs"] mod open_fn;`: pyo3's #[pyfunction] generates a hidden
module with the same name as the function, so a literal `mod open;` plus
`pub use open::open;` would define `open` twice in the type namespace.
