+++ bindings/python/src/database/mod.rs
@@ -35,13 +35,23 @@ use tokio::sync::Mutex as AsyncMutex;
 use pyo3::exceptions::{PyRuntimeError, PyValueError};
 use pyo3::prelude::*;
 
-use crate::config::{config_for_open, StreamingConfig};
+use crate::config::StreamingConfig;
 use crate::error::{runtime_init_to_py_err, to_py_err};
 use crate::prepared::PreparedStatement;
 use crate::result::{QueryResult, StreamingIterator};
 use crate::runtime::block_on;
 use crate::stats::DatabaseStats;
-use crate::write::{MaintenanceReport, PyWriteEngine, WriteStats};
+use crate::write::PyWriteEngine;
+
+// Submodule declarations sit AFTER the `use` block; `open` is reached under an
+// alias because pyo3's `#[pyfunction]` generates a hidden module with the SAME
+// name as the function, so `mod open;` + `pub use open::open;` would define
+// `open` twice in the type namespace.
+#[path = "open.rs"]
+mod open_fn;
+mod write_methods;
+
+pub use open_fn::open;
 
 /// A CQLite database handle.
 ///
@@ -112,41 +122,6 @@ impl Database {
     /// neither is set (the span keeps its natural parent).
     fn resolve_traceparent<'a>(&'a self, per_call: Option<&'a str>) -> Option<&'a str> {
         per_call.or(self.default_traceparent.as_deref())
-    }
-
-    /// The write engine handle, or a typed error if the database is read-only.
-    /// Replaces the `require_writable()? … .expect(Some)` two-step. Returns the
-    /// `&Arc<..>` so callers can `Arc::clone` a `Send` handle into an
-    /// `allow_threads` closure (issue #1444).
-    fn writable_engine(&self) -> PyResult<&Arc<AsyncMutex<PyWriteEngine>>> {
-        self.write_engine.as_ref().ok_or_else(|| {
-            PyRuntimeError::new_err(
-                "Database is read-only. Open with writable=True and write_dir=<path> \
-                 to enable write operations.",
-            )
-        })
-    }
-
-    /// Run `f` with the write-engine guard held, releasing the GIL for the
-    /// duration (issue #1444). The `tokio::sync::Mutex` guard is `Send`, so the
-    /// blocking write executes inside `py.allow_threads` while the lock still
-    /// serializes writes (single-writer: no two flushes/DML overlap).
-    ///
-    /// The guard is taken with `blocking_lock()` on the (synchronous) Python
-    /// calling thread — NOT inside a `block_on`. `PyWriteEngine::flush`/`execute`
-    /// bridge their own async work via `block_on`; acquiring the lock via
-    /// `.lock().await` here would make this thread a runtime driver and nest
-    /// `block_on`s ("cannot start a runtime from within a runtime").
-    fn with_write_engine<T, F>(&self, py: Python<'_>, f: F) -> PyResult<T>
-    where
-        F: FnOnce(&mut PyWriteEngine) -> PyResult<T> + Send,
-        T: Send,
-    {
-        let engine = Arc::clone(self.writable_engine()?);
-        py.allow_threads(move || {
-            let mut guard = engine.blocking_lock();
-            f(&mut guard)
-        })
     }
 
     /// Detect DML statements (INSERT / UPDATE / DELETE / BEGIN BATCH).
@@ -610,346 +585,6 @@ impl Database {
 
         DatabaseStats::from_core(py, core_stats)
     }
-
-    // -----------------------------------------------------------------------
-    // Write API – available only when `writable=True` at open time
-    // -----------------------------------------------------------------------
-
-    /// Flush the memtable to a new SSTable.
-    ///
-    /// Forces a flush of all in-memory writes to disk.  Returns the absolute
-    /// path to the newly written `Data.db` file, or an empty string if the
-    /// memtable was empty (nothing to flush).
-    ///
-    /// Requires the database to have been opened with `writable=True`.
-    ///
-    /// # Returns
-    ///
-    /// Absolute path string of the flushed SSTable Data.db file, or `""` if
-    /// the memtable was empty.
-    ///
-    /// # Raises
-    ///
-    /// * `RuntimeError` – If database is closed or opened in read-only mode
-    /// * `CqliteError`  – If the flush fails (e.g. I/O error)
-    ///
-    /// # Example
-    ///
-    /// ```python
-    /// db.execute("INSERT INTO t (id) VALUES (1)")
-    /// path = db.flush_run()
-    /// assert Path(path).exists()
-    /// ```
-    pub fn flush_run(&self, py: Python<'_>) -> PyResult<String> {
-        self.ensure_open()?;
-        // Flush is substantial I/O (WAL fsync + SSTable materialization). The
-        // `Send` tokio guard lets it run under `allow_threads`, so a large-memtable
-        // flush no longer freezes every other Python thread (issue #1444).
-        self.with_write_engine(py, |e| e.flush().map_err(to_py_err))
-    }
-
-    /// Perform one incremental maintenance step (compaction).
-    ///
-    /// Runs background compaction work within the given time budget.  Can be
-    /// called repeatedly to make incremental progress.  The actual time spent
-    /// may exceed `budget_ms` by up to ~10% (one partition is always processed
-    /// to guarantee forward progress).
-    ///
-    /// Requires the database to have been opened with `writable=True`.
-    ///
-    /// # Arguments
-    ///
-    /// * `budget_ms` – Maximum milliseconds to spend in this call
-    ///
-    /// # Returns
-    ///
-    /// A `MaintenanceReport` with timing, merge statistics, and a
-    /// `pending_compaction` flag.
-    ///
-    /// # Raises
-    ///
-    /// * `RuntimeError` – If database is closed or opened in read-only mode
-    /// * `CqliteError`  – If maintenance fails
-    ///
-    /// # Example
-    ///
-    /// ```python
-    /// report = db.maintenance_step(budget_ms=100)
-    /// print(f"Merged {report.rows_merged} rows in {report.time_spent_ms:.1f} ms")
-    /// ```
-    pub fn maintenance_step(&self, py: Python<'_>, budget_ms: u64) -> PyResult<MaintenanceReport> {
-        self.ensure_open()?;
-        let budget = std::time::Duration::from_millis(budget_ms);
-        // GIL released for the duration; the guard serializes against flush/DML.
-        let report =
-            self.with_write_engine(py, move |e| e.maintenance_step(budget).map_err(to_py_err))?;
-        Ok(MaintenanceReport::from_core(report))
-    }
-
-    /// Current write engine statistics.
-    ///
-    /// Returns a snapshot of memtable occupancy, WAL size, and L0 SSTable count.
-    ///
-    /// Requires the database to have been opened with `writable=True`.
-    ///
-    /// # Returns
-    ///
-    /// A `WriteStats` object.
-    ///
-    /// # Raises
-    ///
-    /// * `RuntimeError` – If database is closed or opened in read-only mode
-    ///
-    /// # Example
-    ///
-    /// ```python
-    /// stats = db.write_stats
-    /// print(f"Memtable: {stats.memtable_size} bytes, {stats.memtable_rows} rows")
-    /// ```
-    #[getter]
-    pub fn write_stats(&self, py: Python<'_>) -> PyResult<WriteStats> {
-        self.ensure_open()?;
-        // Same `Send`-guard path so acquiring the lock never blocks the GIL
-        // against a concurrent flush/DML holding it (issue #1444).
-        self.with_write_engine(py, |e| Ok(e.write_stats()))
-    }
-}
-
-// -----------------------------------------------------------------------
-// Internal helpers (not exposed to Python)
-// -----------------------------------------------------------------------
-
-impl Database {
-    /// Route a DML statement to the write engine.
-    fn execute_dml(&self, py: Python<'_>, query: &str) -> PyResult<QueryResult> {
-        use std::time::Instant;
-
-        let query_owned = query.to_string();
-        let t0 = Instant::now();
-
-        // Release the GIL for the write (issue #1620, #1444). `execute()` routes
-        // through `WriteEngine::execute_flushing`, which performs a REAL async
-        // SSTable flush (disk I/O) once the memtable crosses the flush threshold;
-        // the `Send` tokio guard runs it entirely inside `allow_threads`.
-        let rows_affected =
-            self.with_write_engine(py, move |e| e.execute(&query_owned).map_err(to_py_err))?;
-
-        let elapsed_ms = t0.elapsed().as_millis() as u64;
-
-        QueryResult::from_write(py, rows_affected, elapsed_ms)
-    }
-}
-
-// -----------------------------------------------------------------------
-// open() function
-// -----------------------------------------------------------------------
-
-/// Open a CQLite database.
-///
-/// Opens a database at the specified path, optionally loading a schema file
-/// and applying custom configuration.
-///
-/// # Arguments
-///
-/// * `path`       – Path to the data directory containing SSTables
-/// * `schema`     – Optional path to a CQL schema file (.cql)
-/// * `config`     – Optional configuration (dict, JSON string, or preset name)
-/// * `writable`   – Enable write support (INSERT / UPDATE / DELETE).  Default `False`.
-/// * `write_dir`  – Directory for WAL and flushed SSTables.  Required when
-///                  `writable=True`.  Created automatically if it doesn't exist.
-///
-/// # Caveat: write_dir is not file-locked
-///
-/// Only one `Database` instance should hold a given `write_dir` at a time.
-/// Concurrent instances sharing the same directory are **not** protected by
-/// file locks and will corrupt each other's WAL and SSTable files.  File
-/// locking is planned for a future release (see issue #485).
-///
-/// # Returns
-///
-/// A new Database instance.
-///
-/// # Raises
-///
-/// * `IOError`     – If the path doesn't exist or is inaccessible
-/// * `SchemaError` – If schema parsing fails
-/// * `ValueError`  – If configuration is invalid, or `writable=True` but
-///                   `write_dir` or `schema` was not provided
-///
-/// # Examples
-///
-/// ```python
-/// # Basic read-only open
-/// db = cqlite.open("/path/to/sstables")
-///
-/// # With schema file
-/// db = cqlite.open("/path/to/sstables", schema="/path/to/schema.cql")
-///
-/// # With write support
-/// db = cqlite.open(
-///     "/path/to/sstables",
-///     schema="/path/to/schema.cql",
-///     writable=True,
-///     write_dir="/tmp/cqlite-write",
-/// )
-///
-/// # Using context manager
-/// with cqlite.open("/path/to/sstables") as db:
-///     pass
-/// ```
-#[pyfunction]
-#[pyo3(signature = (path, *, schema=None, config=None, writable=false, write_dir=None, flush_threshold=None, otel_config=None, traceparent=None))]
-pub fn open(
-    py: Python<'_>,
-    path: PathBuf,
-    schema: Option<PathBuf>,
-    config: Option<&Bound<'_, PyAny>>,
-    writable: bool,
-    write_dir: Option<PathBuf>,
-    flush_threshold: Option<u64>,
-    otel_config: Option<&Bound<'_, PyAny>>,
-    traceparent: Option<String>,
-) -> PyResult<Database> {
-    // Initialise observability once per process (issue #1039). The config is
-    // resolved from the optional `otel_config` dict layered over the
-    // `CQLITE_OTEL_*` environment. The first `open` wins; later opens reuse the
-    // installed exporters. A bad exporter config never blocks the open — it
-    // simply yields no telemetry.
-    let obs_cfg = crate::observability::config_from_py(py, otel_config)?;
-    crate::observability::ensure_initialized(obs_cfg);
-
-    // Validate the config-INDEPENDENT half of the flush threshold upfront (#1620),
-    // regardless of `writable`, matching Node: a `0` threshold would make
-    // `should_flush(0)` true after every write. The CEILING half depends on the
-    // caller's `memtable_hard_limit`, so it lives beside the fold below.
-    if let Some(v) = flush_threshold {
-        if v < 1 {
-            return Err(PyValueError::new_err(
-                "flush_threshold must be at least 1 byte",
-            ));
-        }
-    }
-
-    // Validate writable-mode requirements upfront before any I/O.
-    if writable {
-        if write_dir.is_none() {
-            return Err(PyValueError::new_err(
-                "write_dir is required when writable=True. \
-                 Example: cqlite.open(path, writable=True, write_dir='/tmp/cqlite-writes')",
-            ));
-        }
-        if schema.is_none() {
-            return Err(PyValueError::new_err(
-                "schema is required when writable=True so that the write engine \
-                 can resolve column types.",
-            ));
-        }
-    }
-
-    // Assembled in one place (`config::config_for_open`): parse, fold
-    // `flush_threshold`, validate the MERGED result. `Config.storage` is
-    // therefore authoritative for the Python write path rather than decorative
-    // (#1619, #1697); the ONE bridge `WriteEngineConfig::from_config` translates
-    // it below.
-    let core_config = config_for_open(py, config, flush_threshold)?;
-    let write_engine_public_config = core_config.clone();
-
-    // Open the read-side database and capture the schema registry when present.
-    // We always use ingestion when a schema file is provided because that path
-    // performs the SSTable discovery and populates the schema registry.
-    let (db, schema_registry_opt) = if let Some(schema_path) = schema.clone() {
-        let ingestion_config = cqlite_core::ingestion::IngestionConfig {
-            schema_paths: vec![schema_path],
-            data_dir: path.clone(),
-            version_hint: None,
-            core_config,
-            table_directory_filter: None,
-        };
-
-        py.allow_threads(|| {
-            block_on(async {
-                let result = cqlite_core::ingestion::ingest(ingestion_config).await?;
-                let registry = result.schema_registry;
-                Ok::<_, cqlite_core::Error>((result.database, Some(registry)))
-            })
-        })
-        .map_err(runtime_init_to_py_err)?
-        .map_err(to_py_err)?
-    } else {
-        let db = py
-            .allow_threads(|| block_on(cqlite_core::Database::open(&path, core_config)))
-            .map_err(runtime_init_to_py_err)?
-            .map_err(to_py_err)?;
-        (db, None)
-    };
-
-    // Build WriteEngine when writable=True.
-    let write_engine: Option<Arc<AsyncMutex<PyWriteEngine>>> = if writable {
-        let wd = write_dir.expect("validated above");
-        let schema_path_display = schema.clone().expect("validated above");
-
-        // Retrieve the first TableSchema from the registry loaded during ingestion.
-        let table_schema = py
-            .allow_threads(|| {
-                block_on(async {
-                    let registry = schema_registry_opt.ok_or_else(|| {
-                        cqlite_core::Error::Schema(
-                            "Internal error: schema registry unavailable. \
-                             This should not happen when schema= is provided."
-                                .to_string(),
-                        )
-                    })?;
-
-                    let schemas = registry
-                        .read()
-                        .await
-                        .list_schemas(None)
-                        .await
-                        .map_err(|e| {
-                            cqlite_core::Error::Schema(format!("Failed to list schemas: {}", e))
-                        })?;
-
-                    schemas.into_iter().next().ok_or_else(|| {
-                        cqlite_core::Error::Schema(format!(
-                            "No table schema found in {:?}. \
-                             Verify the schema file contains at least one CREATE TABLE statement.",
-                            schema_path_display
-                        ))
-                    })
-                })
-            })
-            .map_err(runtime_init_to_py_err)?
-            .map_err(to_py_err)?;
-
-        // Via the single bridge (#1697).
-        let engine_config = cqlite_core::storage::write_engine::WriteEngineConfig::from_config(
-            &write_engine_public_config,
-            wd.join("data"),
-            wd.join("wal"),
-            table_schema,
-        );
-
-        let engine = cqlite_core::storage::write_engine::WriteEngine::new(engine_config)
-            .map_err(to_py_err)?;
-
-        Some(Arc::new(AsyncMutex::new(PyWriteEngine::new(engine))))
-    } else {
-        // Silence unused-variable warning
-        let _ = (
-            write_dir,
-            schema_registry_opt,
-            write_engine_public_config,
-            flush_threshold,
-        );
-        None
-    };
-
-    Ok(Database {
-        inner: Arc::new(db),
-        closed: Arc::new(AtomicBool::new(false)),
-        write_engine,
-        default_traceparent: traceparent.filter(|s| !s.trim().is_empty()),
-    })
 }
 
 /// Register database types with the Python module.
